VULCAN-18 Rename com.liferay.vulcan.functions to function
