+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/function/EnneaFunction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.vulcan.functions;
+package com.liferay.vulcan.function;
 
 import java.util.Objects;
 import java.util.function.Function;
@@ -21,16 +21,16 @@ import java.util.function.Function;
  * @author Alejandro Hernández
  */
 @FunctionalInterface
-public interface DecaFunction<A, B, C, D, E, F, G, H, I, J, R> {
+public interface EnneaFunction<A, B, C, D, E, F, G, H, I, R> {
 
-	public default <V> DecaFunction<A, B, C, D, E, F, G, H, I, J, V> andThen(
+	public default <V> EnneaFunction<A, B, C, D, E, F, G, H, I, V> andThen(
 		Function<? super R, ? extends V> after) {
 
 		Objects.requireNonNull(after);
-		return (A a, B b, C c, D d, E e, F f, G g, H h, I i, J j) ->
-			after.apply(apply(a, b, c, d, e, f, g, h, i, j));
+		return (A a, B b, C c, D d, E e, F f, G g, H h, I i) -> after.apply(
+			apply(a, b, c, d, e, f, g, h, i));
 	}
 
-	public R apply(A a, B b, C c, D d, E e, F f, G g, H h, I i, J j);
+	public R apply(A a, B b, C c, D d, E e, F f, G g, H h, I i);
 
 }
