+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/function/PentaFunction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.vulcan.functions;
+package com.liferay.vulcan.function;
 
 import java.util.Objects;
 import java.util.function.Function;
@@ -21,15 +21,15 @@ import java.util.function.Function;
  * @author Alejandro Hernández
  */
 @FunctionalInterface
-public interface TetraFunction<A, B, C, D, R> {
+public interface PentaFunction<A, B, C, D, E, R> {
 
-	public default <V> TetraFunction<A, B, C, D, V> andThen(
+	public default <V> PentaFunction<A, B, C, D, E, V> andThen(
 		Function<? super R, ? extends V> after) {
 
 		Objects.requireNonNull(after);
-		return (A a, B b, C c, D d) -> after.apply(apply(a, b, c, d));
+		return (A a, B b, C c, D d, E e) -> after.apply(apply(a, b, c, d, e));
 	}
 
-	public R apply(A a, B b, C c, D d);
+	public R apply(A a, B b, C c, D d, E e);
 
 }
