+++ modules/apps/foundation/vulcan/vulcan-api/src/main/java/com/liferay/vulcan/function/TriFunction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.vulcan.functions;
+package com.liferay.vulcan.function;
 
 import java.util.Objects;
 import java.util.function.Function;
@@ -21,16 +21,15 @@ import java.util.function.Function;
  * @author Alejandro Hernández
  */
 @FunctionalInterface
-public interface EnneaFunction<A, B, C, D, E, F, G, H, I, R> {
+public interface TriFunction<A, B, C, R> {
 
-	public default <V> EnneaFunction<A, B, C, D, E, F, G, H, I, V> andThen(
+	public default <V> TriFunction<A, B, C, V> andThen(
 		Function<? super R, ? extends V> after) {
 
 		Objects.requireNonNull(after);
-		return (A a, B b, C c, D d, E e, F f, G g, H h, I i) -> after.apply(
-			apply(a, b, c, d, e, f, g, h, i));
+		return (A a, B b, C c) -> after.apply(apply(a, b, c));
 	}
 
-	public R apply(A a, B b, C c, D d, E e, F f, G g, H h, I i);
+	public R apply(A a, B b, C c);
 
 }
