+++ modules/apps/foundation/vulcan/vulcan-wiring-osgi/src/main/java/com/liferay/vulcan/wiring/osgi/internal/RoutesBuilderImpl.java
@@ -16,14 +16,14 @@ package com.liferay.vulcan.wiring.osgi.internal;
 
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.vulcan.error.VulcanDeveloperError.MustHaveProvider;
-import com.liferay.vulcan.functions.DecaFunction;
-import com.liferay.vulcan.functions.EnneaFunction;
-import com.liferay.vulcan.functions.HeptaFunction;
-import com.liferay.vulcan.functions.HexaFunction;
-import com.liferay.vulcan.functions.OctaFunction;
-import com.liferay.vulcan.functions.PentaFunction;
-import com.liferay.vulcan.functions.TetraFunction;
-import com.liferay.vulcan.functions.TriFunction;
+import com.liferay.vulcan.function.DecaFunction;
+import com.liferay.vulcan.function.EnneaFunction;
+import com.liferay.vulcan.function.HeptaFunction;
+import com.liferay.vulcan.function.HexaFunction;
+import com.liferay.vulcan.function.OctaFunction;
+import com.liferay.vulcan.function.PentaFunction;
+import com.liferay.vulcan.function.TetraFunction;
+import com.liferay.vulcan.function.TriFunction;
 import com.liferay.vulcan.pagination.PageItems;
 import com.liferay.vulcan.pagination.Pagination;
 import com.liferay.vulcan.representor.Routes;
